+++ R/zzz.R
@@ -5,15 +5,11 @@
 
   on_package_load("vctrs", {
     register_s3_method("vctrs", "vec_proxy", "Period")
-    register_s3_method("vctrs", "vec_proxy_compare", "Period")
-    register_s3_method("vctrs", "vec_proxy_equal", "Period")
     register_s3_method("vctrs", "vec_restore", "Period")
     register_s3_method("vctrs", "vec_ptype2", "Period.Period")
     register_s3_method("vctrs", "vec_cast", "Period.Period")
 
     register_s3_method("vctrs", "vec_proxy", "Duration")
-    register_s3_method("vctrs", "vec_proxy_compare", "Duration")
-    register_s3_method("vctrs", "vec_proxy_equal", "Duration")
     register_s3_method("vctrs", "vec_restore", "Duration")
     register_s3_method("vctrs", "vec_ptype2", "Duration.Duration")
     register_s3_method("vctrs", "vec_ptype2", "Duration.difftime")
@@ -23,8 +19,6 @@
     register_s3_method("vctrs", "vec_cast", "difftime.Duration")
 
     register_s3_method("vctrs", "vec_proxy", "Interval")
-    register_s3_method("vctrs", "vec_proxy_compare", "Interval")
-    register_s3_method("vctrs", "vec_proxy_equal", "Interval")
     register_s3_method("vctrs", "vec_restore", "Interval")
     register_s3_method("vctrs", "vec_ptype2", "Interval.Interval")
     register_s3_method("vctrs", "vec_cast", "Interval.Interval")